requirements.txt
fastapi
uvicorn
pymongo
pydantic
pydantic-settings
httpx
requests
spacy
nltk
numpy
scikit-learn
sentence-transformers
plantuml
app-store-web-scraper
google-play-scraper
typing-extensions
pytest
newspaper4k
pygooglenews
lxml_html_clean
googlenewsdecoder
